backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
email-validator>=2.2.0
httpx>=0.27.0
openai>=1.30.0
langchain>=0.3.0
aiofiles>=23.2.1
python-multipart>=0.0.9

